+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2023-11-05
+
 ### Fixed
 * no inflection **required**
 * bug with lexicon-less extraction
@@ -40,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/fmatter/unboxer/compare/v0.0.2...HEAD
+[Unreleased]: https://github.com/fmatter/unboxer/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/fmatter/unboxer/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/fmatter/unboxer/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/fmatter/unboxer/compare/v0.0.1...v0.0.1